Multi-signal Support
Extended Terminus to be able to listen for multiple signals and trigger the cleanup and shutdown mechanisms on any of the signals bound to it.

This was achieved by refactoring the constructor to accept an array of signals as strings. Once we have an instance, we call the `decorateWithSignalHandler` method for each of the signals in the array.