Support non-promise function options
These functions don't really _need_ to be async but currently the types
require that they are. For example, currently this is an invalid option:

```typescript
onSignal: () => logger.info("signal received")
```

and instead you'd have to write:

```typescript
onSignal: async () => logger.info("signal received")
```

which might trigger a linter error such as typescript-eslint's
`Async method 'onSignal' has no 'await' expression.`

or

```typescript
onSignal: () => {
  logger.info("signal received")
  return Promise.resolve()
}
```

which creates a promise for no reason other than to satisfy the types!

With this PR the types make it clear that you can actually just return
nothing or, in the case of the health checks, return a value without
using async/await or creating a promise.